[AndroidEnv] Move `components/action_type.py` to `//third_party`.
`action_type.py` does not depend on anything in our system so we can already move it to our third-party directory. Notice that I had to create a blueprint file in order for the CL to go through because we need to explicitly declare the files that we own. For now, it's a relatively simple file, but we can put more things there as needed.

PiperOrigin-RevId: 351584986