Strip out get_screencap from adb_controller.
This was only used for real_phone_simulator, which we removed. Removing this function removes the dependency to PIL, which will make the open-sourcing process lighter.

This function can get screenshots from the device using adb. It is very general, but slower than the communication channels that we use for emulator and Vanadium.

PiperOrigin-RevId: 352573217