[AndroidEnv] Add `Task.max_num_steps`.
This CL adds support for a task-specific maximum number of steps. This is
important when evaluating agents on non-time sensitive tasks because the
performance of the emulator varies dramatically across different runs, machines
and cluster load. This CL also adds a note to the config-wide
"maximum_duration_steps" and "maximum_duration_sec". Notice that it may still
be useful to override this setting, but this can be done _outside_ of
AndroidEnv in `env_factories.py` for example.

PiperOrigin-RevId: 354359520