#AndroidEnv New way to read from UNIX pipe.
`EmulatorConsole` was previously "testing" whether the pipe was ready to read
with `select()`, a Unix syscall that tests whether a file descriptor is ready
for a particular event (e.g. read). This was working OK, but it sometimes
caused crashes on borg where file descriptors would be out of range of its
bitmap. Chris and Daniel Rodriguez fixed this in CL/356791187 by using a
different I/O mechanism, `poll()`, which works fine on Linux, but is not
available on Mac OS.

However, testing whether the pipe is ready for reading is not necessary as a
normal `read()` call on a file is blocking and will return once there's
something to be read. What we _really_ want is actually to have a timeout on
the actual `read()` call. This is tremendously hard to achieve cleanly in
Python because if we read the pipe from a separate thread, the thread itself
can get stuck, and while we can detect that from the main thread, Python does
not have a good way to kill the thread or to get the thread unstuck. I tested
spawning new threads once I detect that a thread does not come back with data
within an alotted time, and it works fine, but it "leaks" file descriptors from
each thread opening the pipe. If run forever, this can lead to running out of
file descriptors, which can cause all sorts of problems in the underlying OS.

This CL takes a different approach by setting the `SIGALRM` system call (at the
process level) just before issuing the `read()` call and disabling it after it
returns successfully. In the case where the `read()` call gets stuck forever,
the operating system will interrupt our process and trigger our handler, where
we raise an exception, effectively changing the program execution from blocking
to signal/`Exception` handling, which we can then detect and recover from. The
downside of this approach is that the `SIGALRM` signal is only available on
Unix, and does not work on Windows.

Both spawning threads and using the `SIGALRM` signal work well and I've stress
tested them on Mac OS and Linux for many hours without getting the agent-loop
blocked. We're opting for the version with signals because it is _much_
cleaner, and it works better in the operating systems we're aiming for
initially.

PiperOrigin-RevId: 362535976