Automated g4 rollback of changelist 362535976.
*** Reason for rollback ***

Breaks the environment when ran with dm_env_rpc (See b/183022594).

*** Original change description ***

#AndroidEnv New way to read from UNIX pipe.

`EmulatorConsole` was previously "testing" whether the pipe was ready to read
with `select()`, a Unix syscall that tests whether a file descriptor is ready
for a particular event (e.g. read). This was working OK, but it sometimes
caused crashes on borg where file descriptors would be out of range of its
bitmap. Chris and Daniel Rodriguez fixed this in CL/356791187 by using a
different I/O mechanism, `poll()`, which works fine on Linux, but is not
avai...

***

PiperOrigin-RevId: 363482508