Automated g4 rollback of changelist 363475239.
*** Reason for rollback ***

This will break when we're passing any wrapper_kwargs (which are not included in the default config). I'll try to think of a way to overcome this.

*** Original change description ***

[AndroidEnv] Making sure only correctly passed settings are accepted.

These lines will raise an error if the user tries to pass a settings item which does not exist in the default dictionary. This would help us avoid frustrating cases when incorrectly passed settings are silently ignored.

For example, if instead of writing settings[‘task_config’][‘max_episode_sec’] = 60 they write settings[‘max_episode_sec’] = 60, unless they are warned they would assume that the default value has...

***

PiperOrigin-RevId: 363496622