#AndroidEnv Add thread-based timeout when reading observations from Emulator.
Instead of issuing a Unix SIGALRM signal before `f.read()`, we now do this
entire logic in a separate thread. The call may still get stuck, but the caller
thread may use its own timeout when reading from a thread-safe queue, allowing
the environment to continue to work. When this timeout happens, a new thread is
created to continue the work.

PiperOrigin-RevId: 364358966