#AndroidEnv Remove `AndroidEnv.step()` time tracking.
We were tracking how much time agents were spending "outside" of `step()`, but we only looked at this once or twice at this metric. The calculation of this metric is not entirely accurate because we forgot to reset the timer in some exit points in `step()`, so given that it clutters the function it's better to remove it.

PiperOrigin-RevId: 365267727