#AndroidEnv: Add event handling to `LogcatThread`.
`LogcatThread` runs in a separate thread, consuming logs asynchronously. It
currently contains quite a bit of unrelated logic (e.g. score, rewards,
episodes) that processes these logs in arbitrary ways, making it quite hard to
maintain and to add features.

This CL adds an observer pattern mechanism such that clients can register
handlers for specific "events" that are triggered upon a specific condition.
Events are currently regular expressions matched against log lines, but it is
possible to further extend this to other signals such as timestamps, process
IDs, thread IDs or other information available in Android logs.

This CL also adds a handy function to wait for events, specific or otherwise,
such that execution is blocked up to an optional timeout. This removes the need
for clients to poll the system while waiting for a message, further reducing
latency and CPU usage. This allows us to remove
`LogcatThread.listen_for_message()` and `LogcatThread.has_received_message()`,
which were rather specific to a single use-case and should not live there.

Further changes will follow this CL to make this one more manageable and easier
to review.

PiperOrigin-RevId: 367455917