Fix `TaskManagerTest`.
There were two test cases `test_setup_steps_skipped_if_no_change()` and
`test_setup_steps_interpreted_if_changed()` which were not indented, thus not
part of `TaskManagerTest`. Our testing infrastructure was ignoring them, but
`pytest`, which is used externally, was failing because of that.

This change fixes the indentation and some errors in the test which were not
detected since they were not run before.

PiperOrigin-RevId: 466173225