Make `pixels` and `orientation` values bounded in `[0,255]`.
Both `pixels` and `orientation` are `np.uint8s`, so all their possible
individual array element values are known beforehand. Setting a finite bound
fixes some typing issues in `GymWrapper` where a `spaces.Box` is created with
`dtype` `np.uint8`, but whose bounds are infinite (which it doesn't support).

In the case of `orientation`, we go one step further by limiting the possible
values to only `[0,1]` since it's a one-hot encoding.

PiperOrigin-RevId: 466680597