Try simulator restarts instead of raising `AdbControllerError`.
Today if an `AdbControllerError` occurs in `AndroidEnv` (e.g. in
`Coordinator._wait_for_device()`), this error bubbles up all the way, usually
crashing the actor. Instead of going to that route right away, we can try to
restart the simulator once more. If the simulator still doesn't boot properly,
then we can expect a `TooManyRestartsError` to be raised.

PiperOrigin-RevId: 471075966