Remove two tests from `task_manager_test.py`.
These two tests supposedly try to check whether `setup_steps` are skipped when
they're the same after a call to `update_task()`. However, `TaskManager` does
not have this functionality, so these two tests are not doing that at all.
Plus, they're accessing a private property that is not part of the API.

PiperOrigin-RevId: 474012322