Remove `pytype` escape hatch in `utils.py`.
`np.float32` is not an instance of `np.dtype`. For example:
`isinstance(np.float32, np.dtype) == False` and
`isinstance(np.float32, np.floating) == False`. In particular, `pytype`
complains with `'dtype' object is not callable [not-callable]` when strict
checking is enabled.
However, from `np.__version__ >= 1.20` there is a new
[`np.typing.DTypeLike`](https://numpy.org/devdocs/reference/typing.html#numpy.typing.DTypeLike)
type which does satisfy these requirements. This allows us to remove the escape
hatch in `utils.py` and to properly annotate the function signature.

PiperOrigin-RevId: 482526105