Avoid calling method on deleted attribute.
In some situations `EmulatorSimulator` may have already been closed, so we
should not try to issue gRPC calls to `self._emulator_stub` if it's been
deleted.

This change modifies that logic to set `self._emulator_stub` to `None` instead
of deleting the attribute. This avoids a more expensive and problematic
`hasattr()` call, and possible `AttributeError` exceptions from being raised
when closing the `EmulatorSimulator` object.

PiperOrigin-RevId: 490762959