Replace `SendBroadcast` with `GenericRequest`.
Currently `SendBroadcast` only supports the `action` argument, which prevents
using other arguments such as extras or specific components (`-n`). There are a
lot of possible arguments (see
https://developer.android.com/studio/command-line/adb#IntentSpec) and we could
potentially create a message to capture all arguments for intents, and even
reuse it when starting an activity, but for now it's better to remove the
restriction of only sending actions, and let the client add as many arguments
as they want with a `GenericRequest`.

PiperOrigin-RevId: 502680791