Remove `task_extras_spec()` usage in `AndroidEnv`.
This method is deprecated and will be removed soon. It was created with the
purpose of specifying a priori all the extras that a task would produce. This
however is not only infeasible (and limiting), but also extremely cumbersome in
non static tasks (or multi, sequential tasks). "Extras" are not to be relied
upon like the RL observation, thus they may or may not be available in all
timesteps.

PiperOrigin-RevId: 520756841