Remove references to `extras_spec` in `tasks_guide.md`.
`Task.extras_spec` is not a reliable mechanism since tasks can "lie" about
their extras (or it can also be incomplete and/or just wrong). It also doesn't
play well with dynamic tasks that change shape and/or data types during their
lifetime, which is especially common in continual learning settings. The field
will eventually be removed from the `Task` protobuf message.

PiperOrigin-RevId: 547189863