Introduce `config_classes.py` to define component's inputs.
This change introduces a new file which will define the input of AndroidEnv's
components as dataclasses. We start with `AdbControllerConfig` since it's a
leaf node, but more will be added over time.

This aligns better with how AndroidEnv is used internally at Google and it
improves the type safety since we avoid untyped dicts.

PiperOrigin-RevId: 588905186