Fix broken `ImageRescaleWrapper` due to PIL's `resize()` change.
PIL 10.4.0's `Image.resize()` behavior causes a breakage in
`ImageRescaleWrapper` due to the way numpy arrays are compared. In particular,
in 10.3.0 the `size` argument was cast to a `tuple`, but not in 10.4.0 (see
[changes](https://github.com/python-pillow/Pillow/compare/10.3.0...10.4.0)).
This has since been
[reinstated](https://github.com/python-pillow/Pillow/commit/936012e861ae1458ca01b7bb3bf2f5cade7405e2),
but it did not make it to the 10.4.0 release.

We can still change our code though, so we now simply pass a `tuple` instead of
an `np.array`.

PiperOrigin-RevId: 652127278