Remove `tmp_dir` arg from `AdbCallParser.__init__()` + file leak fix.
This change removes an unnecessary `tmp_dir` argument from `AdbCallParser`
which already uses `tempfile.NamedTemporaryFile()` which works in all
situations we need. This is important because it simplifies the environment
configuration and what needs to be passed to what during `AndroidEnv`'s
loading, and it'll help to make it more modular.

This change also fixes a bug when installing apps (i.e. `.apk` files) via
`blob` requests where the `.apk` files were allocated and written to, but not
deleted. These files lasted until the (successful) termination of the binary,
and would possibly linger around if terminated abruptly.

PiperOrigin-RevId: 683605810