Move async interaction from `Coordinator` to `BaseSimulator`.
This change adds asynchronous interaction to all simulators (i.e. derived
classes of `BaseSimulator`), making this type of interaction less specific
about our specific usage in the `Coordinator`. This is an ongoing effort to
separate the RL parts in AndroidEnv (mostly `AndroidEnv` and `TaskManager`)
from the more generic parts like `BaseSimulator`. The interface of *derived*
classes changed a little: instead of implementing `get_screenshot()` they now
have to implement `_get_screenshot_impl()`.

This change also contains a small fix to `BaseSimulator.__init__()` which
instead of taking individual params, it now takes a single `SimulatorConfig`,
making its initialization the same as other major components. This requires a
small change to the `super().__init__()` call in derived classes, but it
prepares them for future changes to `SimulatorConfig`.

PiperOrigin-RevId: 684273378