Open source `AccessibilityForwarder`.
This change publishes the source code for `AccessibilityForwarder`, our app
that receives accessibility events and trees and sends them externally via
gRPC. This resolves issue #224.

At the moment we do not have build files, this may come later. We are still
deciding the build system (we'd like to use Bazel since it aligns with our
internal build system), but for the moment this provides valuable information
for people trying to use this, especially with
[AndroidWorld](https://github.com/google-research/android_world).

PiperOrigin-RevId: 686532420