Pass `delete_on_close=False` to `NamedTemporaryFile()`.
The behavior of
[`NamedTemporaryFile()`](https://docs.python.org/3/library/tempfile.html#tempfile.NamedTemporaryFile)
is slightly different on Windows when the file is accessed multiple times in
the same context manager. `delete_on_close=False` allows the file to exist
until the end of the context manager, which is the behavior we want in all
platforms.

Unfortunately this argument was only added in Python `3.12`, so we add a switch
to support earlier versions. We may remove it in the future once `3.11` is EOL.

Please see PR https://github.com/google-deepmind/android_env/pull/265 for
details (and thanks to @NingLi670 for opening the original PR).

PiperOrigin-RevId: 732961399